fix: update env script, scraper function deploys on cloudflare
