+++ src/home/authentication.ts
@@ -4,6 +4,7 @@ import { GitHubUser } from "./github-types";
 import { trackReferralCode } from "./register-referral";
 import { displayGitHubUserInformation } from "./rendering/display-github-user-information";
 import { renderGitHubLoginButton } from "./rendering/render-github-login-button";
+import { startIssueScraper } from "./scraper/issue-scraper";
 // import { issueScraper } from "./scraper/issue-scraper";
 
 export async function authentication() {
@@ -23,7 +24,7 @@ export async function authentication() {
     await displayGitHubUserInformation(gitHubUser);
     // <-- Issue Scraper here -->
     // const supabase = getSupabase();
-    // const githubUserName = gitHubUser.login;
-    //await issueScraper(githubUserName, supabase, accessToken || undefined);
+    const githubUserName = gitHubUser.login;
+    await startIssueScraper(githubUserName);
   }
 }
